break apart secrets for CPL construction

+++ pipeline/bin/converge_cpl.sh
@@ -8,8 +8,15 @@ fi
 
 jenkins_url=$(cat jenkins_url)
 
+if [[ ! -f pipeline/tiers/pipeline_secret.yml ]];
+then
+  echo pipeline/tiers/pipeline_secret.yml file must exist with github creds
+  exit 1
+fi
+
 cat <<END > input.yml
 $(cat pipeline/tiers/pipeline.yml)
+$(cat pipeline/tiers/pipeline_secret.yml)
 jenkins_url: ${jenkins_url}
 END
